fix[api]: narrow model-fit process boundary (QVAC-22631)
