+++ packages/model-fit/package.json
@@ -7,7 +7,7 @@
     "build": "npm run build:ts && npm run build:native",
     "build:native": "bare-make generate && bare-make build && bare-make install",
     "build:ts": "tsc -p tsconfig.build.json",
-    "check:generated": "npm run build:ts && git diff --exit-code -- index.js index.d.ts process.js process.d.ts process-runner.js process-runner.d.ts",
+    "check:generated": "npm run build:ts && git diff --exit-code -- index.js index.d.ts process.js process.d.ts process-internal.js process-internal.d.ts process-runner.js process-runner.d.ts",
     "typecheck": "tsc --noEmit -p tsconfig.json",
     "build:pack": "npm run build:ts && mkdir -p dist && npm pack --pack-destination dist",
     "lint": "npm run lint:js && npm run lint:ts",
@@ -20,7 +20,7 @@
     "test:unit": "npm run test:process && npm run test:node",
     "test:node": "npm run build:ts && node --test test/node/process-path.test.cjs",
     "test:process": "npm run build:ts && bare test/desktop/process.test.js",
-    "test:integration": "npm run build:ts && npm run test:integration:generate && bare test/integration/all.js --exit",
+    "test:integration": "npm run build:ts && npm run test:integration:generate && bare test/integration/all.js --exit && npm run test:process",
     "test:integration:generate": "brittle -r test/integration/all.js test/integration/*.test.js && npm run test:mobile:generate",
     "test:mobile:generate": "bare ./scripts/generate-mobile-integration-tests.js",
     "test:mobile:validate": "node scripts/validate-mobile-tests.js",
@@ -38,6 +38,8 @@
       "index.js",
       "process.js",
       "process.d.ts",
+      "process-internal.js",
+      "process-internal.d.ts",
       "process-runner.js",
       "process-runner.d.ts",
       "eslint.config.mjs"
@@ -49,6 +51,8 @@
     "index.d.ts",
     "process.js",
     "process.d.ts",
+    "process-internal.js",
+    "process-internal.d.ts",
     "process-runner.js",
     "process-runner.d.ts",
     "prebuilds",
