+++ packages/model-fit/src/process.ts
@@ -24,13 +24,6 @@ export type FitProcessResponse =
       }
     }
 
-export interface FitProcessOutcome {
-  response: FitProcessResponse
-  exitCode: 0 | 1 | 2
-}
-
-export type FitProcessFit = (config: FitConfig) => FitResult
-
 function isRecord (value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null && !Array.isArray(value)
 }
@@ -144,28 +137,6 @@ export function encodeFitProcessRequest (config: FitConfig): string {
   return encoded
 }
 
-export function parseFitProcessRequest (value: unknown): FitProcessRequest {
-  if (!isRecord(value)) {
-    throw new TypeError('Fit process request must be an object')
-  }
-  if (value['version'] !== FIT_PROCESS_PROTOCOL_VERSION) {
-    throw new TypeError(`Unsupported fit process protocol version: ${String(value['version'])}`)
-  }
-
-  const config = value['config']
-  if (!isRecord(config)) {
-    throw new TypeError('Fit process request config must be an object')
-  }
-  if (typeof config['modelPath'] !== 'string') {
-    throw new TypeError('Fit process request config modelPath must be a string')
-  }
-
-  return {
-    version: FIT_PROCESS_PROTOCOL_VERSION,
-    config: config as unknown as FitConfig
-  }
-}
-
 export function parseFitProcessResponse (value: unknown): FitProcessResponse {
   if (!isRecord(value)) {
     throw new TypeError('Fit process response must be an object')
@@ -204,70 +175,6 @@ export function parseFitProcessResponse (value: unknown): FitProcessResponse {
     }
     default:
       throw new TypeError(`Fit process response status is invalid: ${String(value['status'])}`)
-  }
-}
-
-export function encodeFitProcessResponse (response: FitProcessResponse): string {
-  const encoded = `${JSON.stringify(response)}\n`
-  if (Buffer.byteLength(encoded, 'utf8') > FIT_PROCESS_MAX_RESPONSE_BYTES) {
-    throw new RangeError('Fit process response exceeds 1 MiB')
-  }
-  return encoded
-}
-
-function invocationError (error: unknown): FitProcessResponse {
-  return {
-    version: FIT_PROCESS_PROTOCOL_VERSION,
-    status: 'invocation-error',
-    error: {
-      name: error instanceof Error ? error.name : 'Error',
-      message: error instanceof Error ? error.message : String(error)
-    }
-  }
-}
-
-function boundedInvocationError (error: unknown): FitProcessResponse {
-  const response = invocationError(error)
-  try {
-    encodeFitProcessResponse(response)
-    return response
-  } catch {
-    return invocationError(new RangeError('Fit process response exceeds 1 MiB'))
-  }
-}
-
-export function runFitProcessLine (line: string, fit: FitProcessFit): FitProcessOutcome {
-  if (Buffer.byteLength(line, 'utf8') > FIT_PROCESS_MAX_REQUEST_BYTES) {
-    return {
-      response: boundedInvocationError(new RangeError('Fit process request exceeds 64 KiB')),
-      exitCode: 2
-    }
-  }
-
-  let parsed: unknown
-  try {
-    parsed = JSON.parse(line)
-  } catch (error) {
-    return { response: boundedInvocationError(error), exitCode: 2 }
-  }
-
-  let request: FitProcessRequest
-  try {
-    request = parseFitProcessRequest(parsed)
-  } catch (error) {
-    return { response: boundedInvocationError(error), exitCode: 2 }
-  }
-
-  try {
-    const response: FitProcessResponse = {
-      version: FIT_PROCESS_PROTOCOL_VERSION,
-      status: 'completed',
-      result: fit(request.config)
-    }
-    encodeFitProcessResponse(response)
-    return { response, exitCode: 0 }
-  } catch (error) {
-    return { response: boundedInvocationError(error), exitCode: 1 }
   }
 }
